+++ noodles-gff/src/record.rs
@@ -3,14 +3,14 @@
 pub mod attributes;
 pub(crate) mod fields;
 mod phase;
-mod strand;
 
 use std::{fmt, io};
 
 use noodles_core::Position;
 
 use self::fields::Fields;
-pub use self::{attributes::Attributes, phase::Phase, strand::Strand};
+pub use self::{attributes::Attributes, phase::Phase};
+use crate::feature::record::Strand;
 
 const MISSING: &str = ".";