+++ noodles-gff/CHANGELOG.md
@@ -10,6 +10,8 @@
 
   * gff: Move `RecordBuf` to `feature` module.
 
+  * gff/record: Move `Strand` under `feature::record`.
+
 ## 0.45.0 - 2025-03-08
 
 ### Changed